# backend/docs/swagger.yaml
basePath: /
definitions:
...
  github_com_StephenQiu30_lanverse_backend_internal_media_application.AssetSummary:
    properties:
      byte_size:
        type: integer
      duration_ms:
        type: integer
      file_name:
        type: string
      height:
        type: integer
      id:
        type: string
      kind:
        type: string
      mime_type:
        type: string
      project_id:
        type: string
      revision:
        type: integer
      width:
        type: integer
    type: object
  github_com_StephenQiu30_lanverse_backend_internal_media_application.LibraryAssetSummary:
    properties:
      byte_size:
        type: integer
      duration_ms:
        type: integer
        x-nullable: true
      file_name:
        type: string
      height:
        type: integer
        x-nullable: true
      id:
        type: string
      kind:
        type: string
      mime_type:
        type: string
      revision:
        type: integer
      width:
        type: integer
        x-nullable: true
    type: object
...
  github_com_StephenQiu30_lanverse_backend_internal_media_application.PersonalUploadResult:
    properties:
      asset:
        $ref: '#/definitions/github_com_StephenQiu30_lanverse_backend_internal_media_application.LibraryAssetSummary'
      duplicate_of:
        type: string
        x-nullable: true
    type: object
...
  github_com_StephenQiu30_lanverse_backend_internal_media_application.UploadResult:
    properties:
      asset:
        $ref: '#/definitions/github_com_StephenQiu30_lanverse_backend_internal_media_application.AssetSummary'
      duplicate_of:
        type: string
    type: object
...
  github_com_StephenQiu30_lanverse_backend_internal_platform_httpapi.Problem:
    properties:
      code:
        type: string
      detail:
        type: string
      meta:
        additionalProperties: {}
        type: object
      request_id:
        type: string
      status:
        type: integer
      title:
        type: string
      type:
        type: string
    type: object
...
info:
  contact: {}
  description: Single workspace project query, resource canvas commands and authorized
    media previews.
  title: Lanverse API
  version: "0.1"
paths:
...
  /api/media/library/uploads:
    post:
      consumes:
      - multipart/form-data
      operationId: uploadPersonalMediaAsset
      parameters:
      - description: UUID永久幂等键
        in: header
        name: Idempotency-Key
        required: true
        type: string
      - description: 与项目媒体相同的真实内容/容量门禁，单文件
        in: formData
        name: file
        required: true
        type: file
      - description: 已确认内容和使用权限且不含需要授权的真人
        in: formData
        name: local_review_confirmed
        required: true
        type: boolean
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_StephenQiu30_lanverse_backend_internal_media_application.PersonalUploadResult'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_StephenQiu30_lanverse_backend_internal_platform_httpapi.Problem'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_StephenQiu30_lanverse_backend_internal_platform_httpapi.Problem'
        "413":
          description: Request Entity Too Large
          schema:
            $ref: '#/definitions/github_com_StephenQiu30_lanverse_backend_internal_platform_httpapi.Problem'
        "415":
          description: Unsupported Media Type
          schema:
            $ref: '#/definitions/github_com_StephenQiu30_lanverse_backend_internal_platform_httpapi.Problem'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/github_com_StephenQiu30_lanverse_backend_internal_platform_httpapi.Problem'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/github_com_StephenQiu30_lanverse_backend_internal_platform_httpapi.Problem'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/github_com_StephenQiu30_lanverse_backend_internal_platform_httpapi.Problem'
      summary: 上传人工确认的个人素材
      tags:
      - media
...
  /api/projects/{pid}/media/uploads:
    post:
      consumes:
      - multipart/form-data
      operationId: uploadMediaAsset
      parameters:
      - description: 项目UUID
        in: path
        name: pid
        required: true
        type: string
      - description: UUID幂等键
        in: header
        name: Idempotency-Key
        required: true
        type: string
      - description: 图片20MiB，视频500MiB且60秒（静音VP8/VP9 WebM实际转为MP4），音频100MiB，包内GLB2模型64MiB，TXT/DOCX原件20MiB；图片/视频宽高均≤8192且≤4000万像素；按真实内容验证
        in: formData
        name: file
        required: true
        type: file
      - description: 已确认本地内容、素材使用权限且不含需要授权的真人素材
        in: formData
        name: local_review_confirmed
        required: true
        type: boolean
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_StephenQiu30_lanverse_backend_internal_media_application.UploadResult'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_StephenQiu30_lanverse_backend_internal_platform_httpapi.Problem'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_StephenQiu30_lanverse_backend_internal_platform_httpapi.Problem'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_StephenQiu30_lanverse_backend_internal_platform_httpapi.Problem'
        "413":
          description: Request Entity Too Large
          schema:
            $ref: '#/definitions/github_com_StephenQiu30_lanverse_backend_internal_platform_httpapi.Problem'
        "415":
          description: Unsupported Media Type
          schema:
            $ref: '#/definitions/github_com_StephenQiu30_lanverse_backend_internal_platform_httpapi.Problem'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/github_com_StephenQiu30_lanverse_backend_internal_platform_httpapi.Problem'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/github_com_StephenQiu30_lanverse_backend_internal_platform_httpapi.Problem'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/github_com_StephenQiu30_lanverse_backend_internal_platform_httpapi.Problem'
      summary: 上传本地已人工确认的媒体
      tags:
      - media
...
swagger: "2.0"